Unify "meta" types into "apis/meta/v1"
